requirements.txt
# Core dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-logging==3.8.0
feedparser==6.0.10
praw==7.7.1
python-dotenv==1.0.0
requests==2.31.0

# Utilities
pytz==2023.3
tenacity==8.2.3

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Cloud Functions specific
functions-framework==3.5.0
